# Fast nn.embeddings()