@chenyuxyz 2024-03-15
we can merge this rewrite if it's faster, but I don't think it's O(n) embedding. The reduce still loop through vocabs