@Lama-Thematique 2024-03-16
After verification with the graph this is in fact not O(n).

After looking at tinygrad for a while, I don't think that you can do an O(n) embeddings with the current operations that is in practice faster than the current implementation.

Multiples actual O(n) implementations I tried ended up either hitting the recursion limit when resolving due to reduce operations on the Lazybuffers or just taking minutes to run a case that the current implementation can do in a few ms.

I think this is due to having no way of directly contatenating Tensor/Buffers (or I missed it) and the operations doing something similar (like Tensor.stack) creating very big graphs. When there is a lot of arguments those operations create slow recursive calls. 

Here is the graphs for the Tensor.stack:
```
# Note: obviously this is a toy expample and not representing a real code you would want to use, but this illustrate well  the problem
a = Tensor.randint(10,10)
b = Tensor.randint(10)
c = Tensor.stack((*a.realize(), b.realize()))
c.realize()
```
![image](https://github.com/tinygrad/tinygrad/assets/53999012/c7533328-2407-4ba1-8137-7e7bcaaab18b)

We can see why making it a 1000+ element would hit recursion limit, and if you change the operations for a binary tree like structure with dichotomy to make the graph it larger instead of deeper you just end up with something VERY slow to process (especially on the first run where it tends to take minutes).